'encode' renamed to 'serialize'
